feat: Implement database storage with PostgreSQL

+++ src/main/java/ru/itmo/models/Point.java
@@ -9,13 +9,13 @@ import java.io.Serializable;
 @Getter
 public class Point implements Serializable {
     private double x;
-    private int y;
+    private double y;
     private double r;
     private boolean hit;
     private float processTimeInMs;
     private String requestTime;
 
-    public Point(double x, int y, double r, boolean hit, float processTimeInMs, String requestTime) {
+    public Point(double x, double y, double r, boolean hit, float processTimeInMs, String requestTime) {
         this.x = x;
         this.y = y;
         this.r = r;
